twitchai/requirements.txt
twitchio==2.4.0
aiohttp==3.8.6          # ← КЛЮЧЕВО: НЕ ВЫШЕ 3.8.6
frozenlist==1.4.0       # совместимая версия для aiohttp 3.8
multidict==6.0.4
yarl==1.9.4
requests==2.31.0
PyYAML==6.0.1